[Copter][Debug] Fix double-promotion warning.

+++ ArduCopter/GCS_Mavlink.cpp
@@ -687,7 +687,7 @@ bool GCS_MAVLINK_Copter::try_send_message(enum ap_message id)
             if(copter.camera_mount.get_debug_angles(x, y, z))
             {
                 char str[MAVLINK_MSG_STATUSTEXT_FIELD_TEXT_LEN] {};
-                hal.util->snprintf((char *)str, sizeof(str), "MOUNT: %f\t%f\t%f", x, y, z);
+                hal.util->snprintf((char *)str, sizeof(str), "MOUNT: %f\t%f\t%f", (double)x, (double)y, (double)z);
                 GCS_MAVLINK::send_statustext(MAV_SEVERITY_WARNING, 0xFF, str);
             }
         }
